Fixed headers for io.response()
HttpHeaders are immutable after the body is written to so that is moved to before the body is written to.
The PubNub REST API would respond with 400: no JSON payload and the docs said it should have the Content-Type header. To fix this, if the body data can be decoded to a map using json.decode(), then it is assumed that the Content-Type is a JSON and the appropriate header will be added